oem/arm: support UINT64 boot progress events
Decode Arm BootProgressCode numeric sensor events UINT64 values
and publish the complete eight-byte value.

+++ oem/arm/event/oem_event_manager.hpp
@@ -85,7 +85,7 @@ class OemEventManager
      *
      *  @return PLDM completion code
      */
-    int updateBootProgress(uint32_t presentReading) const;
+    int updateBootProgress(uint64_t presentReading) const;
 
     platform_mc::Manager* manager = nullptr;
 };
